Changed DataType::DOUBLE to FLoat

+++ src/sonic_protocol/cpp_trans_compiler/sonic_protocol_lib/include/sonic_protocol_lib/field_type_def.hpp
@@ -11,7 +11,7 @@ namespace sonic_protocol_lib {
 enum class DataType {
     // Primitive
     INT,
-    DOUBLE,
+    FLOAT,
     STRING,
     BOOL,
